requirements.txt
asgiref==3.6.0
Babel==2.11.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-colorfield==0.8.0
django-crispy-forms==1.14.0
django-money==3.0.0
Pillow==9.3.0
py-moneyed==2.0
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
